+++ World.java
@@ -1,27 +1,27 @@
 public class World //this class is used only to get position of an object in this world
 {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     protected boolean gameOver = false;
     public int gameScore=0;
-    public World(int x, int y)
+    public World(double x, double y)
     {
         this.x = x;
         this.y = y;
     }
-    public int getX()
+    public double getX()
     {
         return x;
     }
-    public int getY()
+    public double getY()
     {
         return y;
     }
-    public void setX(int x)
+    public void setX(double x)
     {
         this.x = x;
     }
-    public void setY(int y)
+    public void setY(double y)
     {
         this.y = y;
     }
